Add several new Handlers:
The Router serves to simply find and invoke a service with the name
found in the MC_TARGET property of the MessageContext.  This guy
will get used a lot in various contexts.

The HTTPActionHandler is a demonstration of transport-specific Handlers.
He just takes the value of the SOAPAction property and puts it into the
MC_TARGET property so that a Router can use it.

BasicHandler is an abstract class to make writing Handlers easier.  It
encapsulates the property database, and stubs for every method
except invoke() (which must be implemented by subclasses).

Changed ErrorHandler + EchoHandler + DebugHandler to extend
BasicHandler.  They're tiny now. :)


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@244597 13f79535-47bb-0310-9956-ffa450edef68