Probably shouldn't close the file just yet!
Where to close the file is actually a bit of a problem...it can't be after
it is read as this is common code on the server which needs the socket to
stick around until the response is complete.  And by the time the client
gets control back, the copy of the InputStream in the Message has been
replaced (possibly several times).  Hmmmm....


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@244763 13f79535-47bb-0310-9956-ffa450edef68