ID/HREF support.  Once again, needs some cleanup, but I've put it through some fairly nasty tests and it works.
Basic pattern is when you start deserializing an element with an
HREF on it, check to see if we have that ID in our table already.
If we do, there should be either an ElementRecorder or a Deserializer
in the table.  In the former case, we play the recorded events to our
Deserializer, and in the latter, we just go grab the value.

If we DON'T already have the ID, we add this deserializer to a
fixup table in the DeserializationContext.  After that, whenever the
SOAPSAXHandler encounters an element with an ID, it checks
against the pending fixups before deciding to just record it.  If there
is a waiting deserializer, it uses that instead.

Deserializers have now changed to "push" their values once they
have them to a set of "targets", which are right now target Fields
of Objects, using the reflection API.  So we can do things like we
do in DataSer (samples/encoding) to fixup a complete structure
made up of parts which themselves need fixups.

This all assumes local refs right now, next step is to abstract out the
HREF dereferencing mechanism to be able to plug in arbitrary handlers
for different types of references (cid:, http:, etc...).

Tests are also a-coming.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@244786 13f79535-47bb-0310-9956-ffa450edef68