Fix potential bug in MessageContext by syncing the targetService string and the actual serviceHandler member.
Rough in support for per-service type mappings by setting the MessageContext
TypeMappingRegistry when we set the target service.  Also change Admin
to parse per-service bean mappings (this is just a P.O.C).

Give AxisFault a String-only constructor for ease-of-use.

Give SOAPHeader a simple constructor for building headers on the fly.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@251948 13f79535-47bb-0310-9956-ffa450edef68