Rudimentary session support.
Session is a base class which provides a simple interface to getting and
setting properties in a session context.  SimpleSession is a trivial
implementation of this with a Hashtable representation.

See next couple of changes for a little more related to this, though Rob
will be doing the work of actually hooking this into a working form.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@244933 13f79535-47bb-0310-9956-ffa450edef68