Serialization / Deserialization for Maps.
Right now serializes both Hashtables and HashMaps, but only Deserializes
to HashMap.

It would be kind of cool in the future to be able to serialize an Interface
rather than inidvidual classes, but the lookup might be a bit expensive
if we have to use reflection and walk the interface list for the object.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@252181 13f79535-47bb-0310-9956-ffa450edef68