Add an optional "basePath" param to the AxisEngine constructor.  This allows us to default the location of the axis.properties and config.xml files.  The AdminServlet sets this to the webapp's WEB-INF directory now.  Default is still ".".
Sketch in some support for limiting access to the Admin service.  Right
now this means not letting any non-local IP address access the service.
Also added an "admin password" property which in the future could be
shipping with a default value, then no admin activity would be allowed
on the server until the user first changed the password.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@252187 13f79535-47bb-0310-9956-ffa450edef68