Remember prefix mappings for recorded elements in a different way.
We now mark the first prefixMapping as the start of the MessageElement's
recorded SAX events, instead of the startElement event.  This should
enable us to correctly serialize recorded elements and remember all the
prefix mappings correctly.

Also add a main() to TestDOM, and test the DOM round-trip in the main()
for TestSerializedRPC.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@252249 13f79535-47bb-0310-9956-ffa450edef68