Change for better WSDL-generation architecture.
Handler now has a generateWSDL() signature, which gets delegated
to chains just like invoke() does.  This allows arbitrary chains of
handlers to collaborate in building up information which ends up
expressed as WSDL.

This stuff is somewhat experimental at the moment, so I labelled the
source immediately prior to the change to make it easy to roll back
just in case.  It does work, and also now generates WSDL inside
the SimpleAxisServer as well.  Check it out, and please comment
on the approach.

Right now all the WSDL is getting created by the JavaProvider.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@245224 13f79535-47bb-0310-9956-ffa450edef68