First cut at EJB provider.
This provider works by looking for a "beanJndiName" option on the
service instead of a "className" one, looking that up in JNDI (it
should be your bean home object), and calling create() on it to get
your actual object.  This is quite simple for now, and only really
works with stateless session beans (no primary key).

To use it, you need four options, which may be set up EITHER on
the EJBProvider itself OR on your service (service specific values
will overrride the provider ones):

jndiURL -> the URL to your JNDI server
jndiContextClass -> the classname of the initial context factory
jndiUser -> username property for the initialContext
jndiPassword -> password for same

An example:

<handler name="EJBProvider" class="org.apache.axis.providers.java.EJBProvider">
  <option name="jndiURL" value="ejipt://localhost:2323"/>
  <option name="jndiContextClass" value="allaire.ejipt.ContextFactory"/>
  <option name="jndiUser" value="fred"/>
  <option name="jndiPassword" value="AllYouNeedIsLove"/>
</handler>
<service name="EJBService" pivot="EJBProvider">
  <option name="beanJndiName" value="myHomeJNDIName"/>
  <option name="allowedMethods" value="*"/>
</service>


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@252299 13f79535-47bb-0310-9956-ffa450edef68