Revamp WSDD, stage 1.
WSDD objects no longer tie tightly to an underlying DOM.  You can
dynamically create them and edit them, and when serialization to XML
is required, we use the SerializationContext (eating our own dog food,
as it were) to do it.  This gets rid of a number of problems, including
duplicate entries and messy files, which were plaguing the earlier
version.

Passes func tests, but still needs more cleanup.  Checking it in to
checkpoint it at this stage.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@252769 13f79535-47bb-0310-9956-ffa450edef68