OK - With Russell's help here's a fix for the functional-tests. I was assuming that any untyped value must be the return of the RPC - when in fact I should have checked for any OUTPUT params first (by name).
Perhaps we could agree to NOT checkin failing testcases since it
leaves people in a state where if they make additional changes
they don't know if their new code is what broke things or if it
is supposed to fail or if someone else broke it.  Ah, the importance
of making sure you don't commit until: ant clean all   works.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@245706 13f79535-47bb-0310-9956-ffa450edef68