If a <complexType> contains a <sequence> group element, the elements must be written out in the prescribed order when serialized.  A <all> group does not dictate an ordering.
However, by the type a java object gets to the BeanSerializer, there is
no way to determine if it was constructed from a <sequence> or <all> group.
To make matters worse, the PropertyDescriptors (from BeanInfo) are in a
'random' order.

I added a sortPropertyDescriptors() utility method to the BeanSerializer to attempt to
sort the PropertyDescriptors according to the order of the 'get' methods.
(The assumption is that the order of the get methods is the same as the
order of the elements in a <sequence>...a good assumption if the class was
generated by an emitter.)  There is little performance penalty since the
PropertyDescriptors are obtained once per class.

(I could have added coded to the BeanDeserializer 'find field by name' loop
to anticipate the next field being deserialized...)

In the process of making this change I found/fixed a NullPointerException in Deserializer.

I also added a sequence test to make verify compliance.


PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@252825 13f79535-47bb-0310-9956-ffa450edef68