Update the JAX-RPC Call API to support some of the properties specified in the spec. Specifically, http.auth.user and http.auth.password are now set correctly in the Call object in the stub. This allows users of the stubs to set username and password. The Call api will set the correct messageContext properties.
In addition, we support the soap.http.soapaction.uri and soap.http.soapaction.use properties.
This is kind of tricky since the default in JAX-RPC is to NOT use the soapaction property unless
the "use" flag is set to true.  I modified the stubs to set the use flag and soapaction if it is non-null.

TBD is support for:  security.auth.subject, soap.operation.style and encodingstyle.namespace.uri
Probably the right thing is to have MessageContext support these properties directly.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@246111 13f79535-47bb-0310-9956-ffa450edef68