Enhance WSDL2Java provide more information on what exactly was generated.
Added a routine to WSDL2Java class which returns an object (GeneratedFileInfo) which contains a list of
the files/classes generated and what their types are (e.g. stub, skeleton, interface, etc.).

This object can be queried for a class name, file name, or entries of a particular type.

This list is used instead of seperate class/filename lists to prevent duplicates.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@253237 13f79535-47bb-0310-9956-ffa450edef68