Get rid of AxisClassLoader, which was trying to wear too many hats. Split the code functionality into JWSClassLoader, a simple ClassLoader extension which allows us to reload JWS classes easily, and ClassCache, a repository for JavaClass objects.
MessageContext.getClassLoader() now always returns the thread's
contextClassLoader EXCEPT when that has been replaced by the
JWSProcessor with a JWSClassLoader.

So... 1) our code is simpler, and 2) changing a JWS file now actually
causes the class to reload on the next hit (this hadn't been working for
some time now).

NOTES:

I've left the AxisClassLoader class for now, although the only thing that
references it is the JUnit test which tests that class.  I plan to delete it
entirely if people are happy with these changes.

I'm a little dubious about the lookups in JavaMethod, which seem to be the
only real reason for having the ClassCache around in the first place.  Do
these really improve performance that much?


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@246368 13f79535-47bb-0310-9956-ffa450edef68