Patch from Dave Dunkin [DaveDu@Attachmate.com] :
The following patch fixes a few cases that aren't mapped correctly:
An identifier where the first character is a mark (e.g. "\u2160foo" maps
to "\u2160Foo").
An identifier where the first character is a valid Java identifier
start, but is puctuation (e.g. "_foo" maps to "foo").

It also handles identifiers that don't contain any characters that are
valid in Java identifiers (e.g. "-").


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@253488 13f79535-47bb-0310-9956-ffa450edef68