Support for <attribute> WSDL (required for interop tests).
WSDL2Java will generate bean fields for XML element attributes if specified
in the WSDL.  It wil also generate a getAttributeElements() static function in
the bean that the bean serializer/deserializer uses.  (The modifications to the
serialization code have already been committed.)
Sample:
public class AccessPoint implements java.io.Serializable {
    private String element;
    private String myAttribute;  // attribute
...
    public static java.util.Vector getAttributeElements() {
        java.util.Vector v = new java.util.Vector();
        v.add("myAttribute");
        return v;
    }
}

Java2WSDL will generate <attribute> in the type schema for beans that have
the getAttributeElements()  method to id which fields should be attributes.
This involved moving the WSDL type generation for beans out of Java2WSDL
and in to the serializer itself.

TBD:
 - we do not handle all forms of attributes and the UDDI WSDL in our test
suite is commented out (this WSDL uses LOTS of attributes).

- The generation of type Schema should live in the serializers for all types.
Currently only Map and Bean have this code.

- For command line Java2WSDL, type serializers are generally not registered,
so we default to the BeanSerializer for everything that isn't built in.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@246470 13f79535-47bb-0310-9956-ffa450edef68