Add missing classes.
These are the very beginnings of an abstract service metadata framework.

The goal here is that these classes (plus ParameterDesc, and with some
fleshing out) become the way the engine thinks about a service, and
contain everything we might need to know about a service with convenient
APIs to get at all the useful information in various parts of the engine/tools.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@246594 13f79535-47bb-0310-9956-ffa450edef68