Implement structured type mapping metadata, part I.
* Define two new classes, TypeDesc and FieldDesc, representing
  metadata about a Java<->XML mapping for a given type.  FieldDesc
  maps a given Java field to/from a particular XML element OR attribute,
  and has two subclasses, ElementDesc and AttributeDesc.  The subclasses
  may be created directly, the FieldDesc root may not.

* Adjust the code in the BeanSerializer/Deserializer and SimpleSerializer/
  Deserializer to look for a static "getTypeDesc()" method on classes to
  obtain metadata instead of looking for the old "getAttributeNames()"
  string array.

* Edit the stub generator to emit the new metadata style.

* Fix up the TestAttributes test to use the new metadata style, test
  more combinations of present + missing metadata, and also add a
  test for a SimpleType with an attribute.

So what does this mean?

The framework is now in place for much smoother data binding, and we
can eventually move towards supporting an xdoclet-type Javadoc
syntax for expressing these kinds of mappings.

We now support attributes which are in other namespaces than the
default.

Next steps:

Get the emitter doing the right thing for XML -> Java name mangling.  In
other words, if we have to change an XML name when writing a stub
from WSDL, we should now be able to keep track of the original XML
name and add it to the metadata instead of relying on the Java -> XML
reverse name mangling to preserve fidelity.

Get the WSDL emitted for attributes to look correct.  I was a little too
tired to finish this tonight.

Attempt to refactor the BeanSerializer/SimpleSerializer and equiv.
deserializers into a common family so they can share all the duplicated
code that currently lives both places.

Clean up a bit more.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@253698 13f79535-47bb-0310-9956-ffa450edef68