Patch from Dave Dunkin [DaveDu@Attachmate.com]
The attached patch addresses a problem with JavaProvider where if a
class is not found while generating the WSDL, the NoClassDefFoundError
makes it all the way out to the servlet engine. Instead, an AxisFault
should be generated that contains the message from the
NoClassDefFoundError. I also fixed a few debug statements.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@246698 13f79535-47bb-0310-9956-ffa450edef68