Improved the Stub assignment and conversion of output parameters.
There are a number of places in JavaStubWriter that
assigned (or returned) the output value(s).  If the
expected type is an array, JavaUtils.convert is invoked
to convert the value to the expected type.  Otherwise
the output value is cast to to the expected type.

This processing fails if an array comes over the wire
and the expected type is a Vector.

The following changes are made to the stub generation:

   First an attempt is made to cast the
   output value to the expected type.
   If an exception is thrown, it is caught and
   an attempt is made to convert the output value
   to the expected type using JavaUtils.convert.

   This slightly improves performance for the
   array case (since most of the time a cast is
   sufficient).  It also makes the stub more
   flexible, allowing conversions in many more
   situations.

   There was a lot of redundant code in JavaStubWriter,
   so I moved most of the processing into a private
   method (writeOutputAssign).  I also cleaned
   up some bad logic and stale comments.

   I changed the arrays testcase so that
   it returns a Vector.  This verifies the
   new stub code.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@253832 13f79535-47bb-0310-9956-ffa450edef68