Added minOccursIs0 property to ElementDecl and changed SchemaUtils to set this property to true if a contained element has maxOccurs="0".
This information is useful for the subtle distinction between nillable="true"
and maxOccurs="0".  Currently the runtime assumes nillable="true" and
sends null values as
  <someItem xsi:nill="true" />

But if maxOccurs="0" is defined for the element, the element should not be
serialized at all.

Glen/Tom: You may want to change JavaBeanHelperWriter to store meta data
for this case, add this information to the meta data model and change
the BeanSerializer.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@253915 13f79535-47bb-0310-9956-ffa450edef68