Change to deal with namespace qualification better and fix several bugs.
Add qualify test to WSDL test suite.
 This tests the form="qualified/unqualified" and
 defaultFormElement="qualified/unqualified" schema attributes.

Fix a bug in WSDDService that was always setting the
encoding style for type mappings to RPC encoded.

In the WSDD operation element:
 - the <elementMapping> tag has been removed.  The code
   to read it is still there, but commented out.
 - A "qname" attribute for the operation was added to provide
   the element name/qname to operation description mapping.

For Anonymous type QNames:
The ">" is retained on all QNames since they are only used
internally.  Put a check in Serialization content to prevent
writing xsi:types that are anonymous types.

Lookup allowed methods by method name instead of XML name.

Add a routine genQNameAttributeString() in toJava.Utils to
generate an attribute value string that is a valid QName,
including any necessary prefix mapping.

In the SymbolTable, set the refType of elements with
anonymous types under them to point to the anonymous
types. Added a setRefType() setter to TypeElement for
this.  Modified JavaTypeWriter to handle the change when
javafying names.

In the ServiceDescription, added a getOperationsByQName() function
which is used to lookup potentially overloaded operations by QName.

Improved the structure of how RPCElements/RPCHandler deals with
operation descriptors.  RPCHandler now contains the operation
instead od RPCElement.  Still some work to be done here, but
we now lookup operaions by QName and deal with overrides much better.

Fix a serious bug in NSStack having to do with prefix mappings.
See comment in getPrefixForURI() for details.

JavaDeployWriter now emits the new qname attribute of <operation>.
This replaces elementMapping subtag.

Fix a bug in Allowed methods that was using the wrong operaion names
(XML instead of Javafied).


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@246864 13f79535-47bb-0310-9956-ffa450edef68