A bug in the WSDDService object was reported by Greg Truty.
He is building a WSDDService on the fly and using the
addOperation code to add WSDDOperations.

The code was failing to add the corresponding OperationDesc
to the ServiceDesc.  This resulted in a number of hard
to find problems during sync processing.

Thanks to Greg for debugging this problem and proposing the following
solution.


PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@246970 13f79535-47bb-0310-9956-ffa450edef68